requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
orjson>=3.9.0
aiofiles>=23.0.0
